improve DynamicEndpointSnitch by using reservoir sampling patch by Pavel Yaskevich; reviewed by Brandon Williams for CASSANDRA-4038

+++ src/java/org/apache/cassandra/locator/ILatencySubscriber.java
@@ -21,5 +21,5 @@ import java.net.InetAddress;
 
 public interface ILatencySubscriber
 {
-    public void receiveTiming(InetAddress address, Double latency);
+    public void receiveTiming(InetAddress address, long latency);
 }
